refactor: move logger code out of fastify
