+++ lib/logger-utils.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const { validateLogger } = require('./logger')
+const { validateLogger } = require('./logger-factory')
 
 /**
  * Utility for creating a child logger with the appropriate bindings, logger factory
