+++ lib/reply.js
@@ -32,7 +32,7 @@ const {
 } = require('./hooks')
 
 const internals = require('./handleRequest')[Symbol.for('internals')]
-const loggerUtils = require('./logger')
+const loggerUtils = require('./logger-factory')
 const now = loggerUtils.now
 const { handleError } = require('./error-handler')
 const { getSchemaSerializer } = require('./schemas')
