+++ test/internals/logger.test.js
@@ -2,7 +2,8 @@
 
 const { test } = require('node:test')
 const Fastify = require('../..')
-const loggerUtils = require('../../lib/logger')
+const loggerUtils = require('../../lib/logger-factory')
+const { serializers } = require('../../lib/logger-pino')
 
 test('time resolution', t => {
   t.plan(2)
@@ -144,7 +145,7 @@ test('The logger should error if both stream and file destination are given', t 
 test('The serializer prevent fails if the request socket is undefined', t => {
   t.plan(1)
 
-  const serialized = loggerUtils.serializers.req({
+  const serialized = serializers.req({
     method: 'GET',
     url: '/',
     socket: undefined,
